+++ core/new-gui/src/app/dashboard/user/component/filters/filters.component.ts
@@ -2,7 +2,7 @@ import { Component, EventEmitter, OnInit, Output } from "@angular/core";
 import { OperatorMetadataService } from "src/app/workspace/service/operator-metadata/operator-metadata.service";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { Observable } from "rxjs";
-import { DashboardProject } from "../../type/dashboard.project";
+import { DashboardProject } from "../../type/dashboard-project.interface";
 import { remove } from "lodash-es";
 import { NotificationService } from "src/app/common/service/notification/notification.service";
 import { UserProjectService } from "../../service/user-project/user-project.service";
