+++ core/new-gui/src/app/dashboard/user/component/user-project/user-project-list-item/user-project-list-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from "@angular/core";
-import { DashboardProject } from "../../../type/dashboard.project";
+import { DashboardProject } from "../../../type/dashboard-project.interface";
 import { UserProjectService } from "../../../service/user-project/user-project.service";
 import { NotificationService } from "src/app/common/service/notification/notification.service";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
