+++ core/new-gui/src/app/dashboard/user/type/dashboard-project.interface.ts
